delete specific files now
